requirements.txt
pandas
python-dwca-reader
langchain
langchain-openai
langchain_huggingface
ipykernel
google-cloud-vision
fuzzywuzzy
plotly
langchain-anthropic
requests
